refactor fulfillment zome in planning DNA into modular crates
only minor visibility code changes
