+++ happs/observation/zomes/fulfillment/code/Cargo.toml
@@ -15,8 +15,12 @@ holochain_wasm_utils = "=0.0.40-alpha1"
 holochain_json_derive = "=0.0.17"
 
 hdk_graph_helpers = { path = "../../../../../lib/hdk_graph_helpers" }
-vf_planning = { path = "../../../../../lib/vf_planning" }
+vf_core = { path = "../../../../../lib/vf_core" }
 hc_zome_rea_economic_event_storage_consts = { path = "../../../../../lib/rea_economic_event/storage_consts" }
+hc_zome_rea_fulfillment_storage_consts = { path = "../../../../../lib/rea_fulfillment/storage_consts" }
+hc_zome_rea_fulfillment_storage = { path = "../../../../../lib/rea_fulfillment/storage" }
+hc_zome_rea_fulfillment_rpc = { path = "../../../../../lib/rea_fulfillment/rpc" }
+hc_zome_rea_fulfillment_lib = { path = "../../../../../lib/rea_fulfillment/lib" }
 
 [lib]
 path = "src/lib.rs"
