+++ happs/observation/zomes/fulfillment/code/src/fulfillment_requests.rs
@@ -3,11 +3,8 @@
  */
 
 use hdk::{
-    // PUBLIC_TOKEN,
     holochain_json_api::{ json::JsonString, error::JsonError },
-    error::ZomeApiResult,
-    error::ZomeApiError,
-    // call,
+    error::{ ZomeApiResult, ZomeApiError },
 };
 use holochain_json_derive::{ DefaultJson };
 use hdk_graph_helpers::{
@@ -23,26 +20,15 @@ use hdk_graph_helpers::{
     },
 };
 
-use vf_planning::type_aliases::{FulfillmentAddress, EventAddress};
+use vf_core::type_aliases::{FulfillmentAddress, EventAddress};
 use hc_zome_rea_economic_event_storage_consts::{
     EVENT_FULFILLS_LINK_TYPE,
     EVENT_FULFILLS_LINK_TAG,
 };
-use vf_planning::identifiers::{
-    FULFILLMENT_BASE_ENTRY_TYPE,
-    FULFILLMENT_INITIAL_ENTRY_LINK_TYPE,
-    FULFILLMENT_ENTRY_TYPE,
-    FULFILLMENT_FULFILLEDBY_LINK_TYPE,
-    FULFILLMENT_FULFILLEDBY_LINK_TAG,
-};
-
-use vf_planning::fulfillment::{
-    Entry,
-    CreateRequest,
-    UpdateRequest,
-    ResponseData as Response,
-    construct_response,
-};
+use hc_zome_rea_fulfillment_storage_consts::*;
+use hc_zome_rea_fulfillment_storage::Entry;
+use hc_zome_rea_fulfillment_rpc::*;
+use hc_zome_rea_fulfillment_lib::construct_response;
 
 #[derive(Serialize, Deserialize, Debug, DefaultJson, Clone)]
 #[serde(rename_all = "camelCase")]
@@ -50,15 +36,15 @@ pub struct QueryParams {
     fulfilled_by: Option<EventAddress>,
 }
 
-pub fn receive_create_fulfillment(fulfillment: CreateRequest) -> ZomeApiResult<Response> {
+pub fn receive_create_fulfillment(fulfillment: CreateRequest) -> ZomeApiResult<ResponseData> {
     handle_create_fulfillment(&fulfillment)
 }
 
-pub fn receive_get_fulfillment(address: FulfillmentAddress) -> ZomeApiResult<Response> {
+pub fn receive_get_fulfillment(address: FulfillmentAddress) -> ZomeApiResult<ResponseData> {
     handle_get_fulfillment(&address)
 }
 
-pub fn receive_update_fulfillment(fulfillment: UpdateRequest) -> ZomeApiResult<Response> {
+pub fn receive_update_fulfillment(fulfillment: UpdateRequest) -> ZomeApiResult<ResponseData> {
     handle_update_fulfillment(&fulfillment)
 }
 
@@ -66,11 +52,11 @@ pub fn receive_delete_fulfillment(address: FulfillmentAddress) -> ZomeApiResult<
     delete_record::<Entry>(&address)
 }
 
-pub fn receive_query_fulfillments(params: QueryParams) -> ZomeApiResult<Vec<Response>> {
+pub fn receive_query_fulfillments(params: QueryParams) -> ZomeApiResult<Vec<ResponseData>> {
     handle_query_fulfillments(&params)
 }
 
-fn handle_create_fulfillment(fulfillment: &CreateRequest) -> ZomeApiResult<Response> {
+fn handle_create_fulfillment(fulfillment: &CreateRequest) -> ZomeApiResult<ResponseData> {
     let (fulfillment_address, entry_resp): (FulfillmentAddress, Entry) = create_record(
         FULFILLMENT_BASE_ENTRY_TYPE, FULFILLMENT_ENTRY_TYPE,
         FULFILLMENT_INITIAL_ENTRY_LINK_TYPE,
@@ -98,19 +84,19 @@ fn handle_create_fulfillment(fulfillment: &CreateRequest) -> ZomeApiResult<Respo
     Ok(construct_response(&fulfillment_address, &entry_resp))
 }
 
-fn handle_update_fulfillment(fulfillment: &UpdateRequest) -> ZomeApiResult<Response> {
+fn handle_update_fulfillment(fulfillment: &UpdateRequest) -> ZomeApiResult<ResponseData> {
     let base_address = fulfillment.get_id();
     let new_entry = update_record(FULFILLMENT_ENTRY_TYPE, base_address, fulfillment)?;
     Ok(construct_response(&base_address, &new_entry))
 }
 
 /// Read an individual fulfillment's details
-fn handle_get_fulfillment(base_address: &FulfillmentAddress) -> ZomeApiResult<Response> {
+fn handle_get_fulfillment(base_address: &FulfillmentAddress) -> ZomeApiResult<ResponseData> {
     let entry = read_record_entry(base_address)?;
     Ok(construct_response(base_address, &entry))
 }
 
-fn handle_query_fulfillments(params: &QueryParams) -> ZomeApiResult<Vec<Response>> {
+fn handle_query_fulfillments(params: &QueryParams) -> ZomeApiResult<Vec<ResponseData>> {
     let mut entries_result: ZomeApiResult<Vec<(FulfillmentAddress, Option<Entry>)>> = Err(ZomeApiError::Internal("No results found".to_string()));
 
     match &params.fulfilled_by {
