+++ lib/rea_economic_event/defs/Cargo.toml
@@ -8,12 +8,12 @@ edition = "2018"
 # :DUPE: hdk-rust-revid
 hdk = "=0.0.40-alpha1"
 
-vf_planning = { path = "../../vf_planning" }
 hc_zome_rea_economic_event_storage_consts = { path = "../storage_consts" }
 hc_zome_rea_economic_event_storage = { path = "../storage" }
 hc_zome_rea_economic_event_lib = { path = "../lib" }
 hc_zome_rea_process_storage_consts = { path = "../../rea_process/storage_consts" }
 hc_zome_rea_satisfaction_storage_consts = { path = "../../rea_satisfaction/storage_consts" }
+hc_zome_rea_fulfillment_storage_consts = { path = "../../rea_fulfillment/storage_consts" }
 
 [lib]
 crate-type = ["lib"]
