+++ lib/rea_economic_event/defs/src/lib.rs
@@ -9,9 +9,7 @@
  */
 use hdk::prelude::*;
 
-use vf_planning::identifiers::{
-    FULFILLMENT_BASE_ENTRY_TYPE,
-};
+use hc_zome_rea_fulfillment_storage_consts::FULFILLMENT_BASE_ENTRY_TYPE;
 use hc_zome_rea_process_storage_consts::PROCESS_BASE_ENTRY_TYPE;
 use hc_zome_rea_satisfaction_storage_consts::SATISFACTION_BASE_ENTRY_TYPE;
 use hc_zome_rea_economic_event_storage_consts::*;
