+++ lib/rea_fulfillment/rpc/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "vf_planning"
+name = "hc_zome_rea_fulfillment_rpc"
 version = "0.1.0"
 authors = ["pospi <pospi@spadgos.com>"]
 edition = "2018"
@@ -8,13 +8,11 @@ edition = "2018"
 serde = "1.0.89"
 serde_json = { version = "=1.0.39", features = ["preserve_order"] }
 serde_derive = "=1.0.89"
-# :DUPE: hdk-rust-revid
-hdk = "=0.0.40-alpha1"
+holochain_json_api = "=0.0.17"
 holochain_json_derive = "=0.0.17"
 
-hdk_graph_helpers = { path = "../hdk_graph_helpers" }
-vf_core = { path = "../vf_core" }
-vf_actions = { path = "../vf_actions" }
+hdk_graph_helpers = { path = "../../hdk_graph_helpers" }
+vf_core = { path = "../../vf_core" }
 
 [lib]
 crate-type = ["lib"]
