+++ lib/rea_fulfillment/storage/src/lib.rs
@@ -1,0 +1,77 @@
+/**
+ * Holo-REA fulfillment zome internal data structures
+ *
+ * Required by the zome itself, and for any DNA-local zomes interacting with its
+ * storage API directly.
+ *
+ * @package Holo-REA
+ */
+extern crate serde;
+#[macro_use]
+extern crate serde_derive;
+extern crate serde_json;
+
+use holochain_json_api::{ json::JsonString, error::JsonError };
+use holochain_json_derive::{ DefaultJson };
+
+use hdk_graph_helpers::{
+    MaybeUndefined,
+    record_interface::Updateable,
+};
+
+use vf_core::{
+    measurement::QuantityValue,
+    type_aliases::{
+        EventAddress,
+        CommitmentAddress,
+    },
+};
+
+use hc_zome_rea_fulfillment_rpc::{ CreateRequest, UpdateRequest };
+
+//---------------- RECORD INTERNALS & VALIDATION ----------------
+
+#[derive(Serialize, Deserialize, Debug, DefaultJson, Clone)]
+pub struct Entry {
+    pub fulfilled_by: EventAddress,
+    pub fulfills: CommitmentAddress,
+    pub resource_quantity: Option<QuantityValue>,
+    pub effort_quantity: Option<QuantityValue>,
+    pub note: Option<String>,
+}
+
+//---------------- CREATE ----------------
+
+/// Pick relevant fields out of I/O record into underlying DHT entry
+impl From<CreateRequest> for Entry {
+    fn from(e: CreateRequest) -> Entry {
+        Entry {
+            fulfilled_by: e.fulfilled_by.into(),
+            fulfills: e.fulfills.into(),
+            resource_quantity: e.resource_quantity.into(),
+            effort_quantity: e.effort_quantity.into(),
+            note: e.note.into(),
+        }
+    }
+}
+
+//---------------- UPDATE ----------------
+
+/// Handles update operations by merging any newly provided fields
+impl Updateable<UpdateRequest> for Entry {
+    fn update_with(&self, e: &UpdateRequest) -> Entry {
+        Entry {
+            fulfilled_by: match &e.fulfilled_by {
+                MaybeUndefined::Some(fulfilled_by) => fulfilled_by.clone(),
+                _ => self.fulfilled_by.clone(),
+            },
+            fulfills: match &e.fulfills {
+                MaybeUndefined::Some(fulfills) => fulfills.clone(),
+                _ => self.fulfills.clone(),
+            },
+            resource_quantity: if e.resource_quantity== MaybeUndefined::Undefined { self.resource_quantity.clone() } else { e.resource_quantity.clone().into() },
+            effort_quantity: if e.effort_quantity== MaybeUndefined::Undefined { self.effort_quantity.clone() } else { e.effort_quantity.clone().into() },
+            note: if e.note== MaybeUndefined::Undefined { self.note.clone() } else { e.note.clone().into() },
+        }
+    }
+}
