+++ lib/rea_fulfillment/storage_consts/src/lib.rs
@@ -1,0 +1,16 @@
+/**
+ * Storage constants for zome entry & link type identifiers
+ *
+ * Used by modules interfacing with the underlying Holochain storage system directly.
+ *
+ * @package Holo-REA
+ */
+pub const FULFILLMENT_BASE_ENTRY_TYPE: &str = "vf_fulfillment_baseurl";
+pub const FULFILLMENT_INITIAL_ENTRY_LINK_TYPE: &str = "vf_fulfillment_entry";
+pub const FULFILLMENT_ENTRY_TYPE: &str = "vf_fulfillment";
+pub const FULFILLMENT_FULFILLS_LINK_TYPE: &str = "vf_fulfillment_fulfills";
+pub const FULFILLMENT_FULFILLS_LINK_TAG: &str = "fulfills";
+pub const FULFILLMENT_FULFILLEDBY_LINK_TYPE: &str = "vf_fulfillment_fulfilled_by";
+pub const FULFILLMENT_FULFILLEDBY_LINK_TAG: &str = "fulfilled_by";
+
+pub const BRIDGED_OBSERVATION_DHT: &str = "vf_observation";
