+++ lib/rea_intent/defs/Cargo.toml
@@ -13,7 +13,6 @@ hdk = "=0.0.40-alpha1"
 holochain_json_derive = "=0.0.17"
 
 hdk_graph_helpers = { path = "../../hdk_graph_helpers" }
-vf_planning = { path = "../../vf_planning" }
 hc_zome_rea_intent_storage = { path = "../storage" }
 hc_zome_rea_intent_storage_consts = { path = "../storage_consts" }
 hc_zome_rea_intent_rpc = { path = "../rpc" }
